requirements.txt
streamlit==1.32.0
google-generativeai==0.3.2
python-dotenv==1.0.0